feat: move generator router code to router plugin
