+++ packages/runtime/plugin-runtime/package.json
@@ -193,11 +193,13 @@
     "@modern-js/plugin": "workspace:*",
     "@modern-js/types": "workspace:*",
     "@modern-js/utils": "workspace:*",
+    "@modern-js/plugin-data-loader": "workspace:*",
     "@modern-js/runtime-utils": "workspace:*",
     "@types/loadable__component": "^5.13.4",
     "@types/react-helmet": "^6.1.2",
     "@types/styled-components": "^5.1.14",
     "esbuild": "0.17.19",
+    "es-module-lexer": "^1.1.0",
     "hoist-non-react-statics": "^3.3.2",
     "invariant": "^2.2.4",
     "react-helmet": "^6.1.0",
