+++ packages/runtime/plugin-runtime/src/router/cli/code/getClientRoutes/getRoutesLegacy.ts
@@ -1,19 +1,17 @@
 import path from 'path';
 import { fs } from '@modern-js/utils';
-import type {
-  Entrypoint,
-  NestedRouteForCli,
-  PageRoute,
-} from '@modern-js/types';
-import { makeLegalIdentifier } from '../makeLegalIdentifier';
+import type { Entrypoint, RouteLegacy } from '@modern-js/types';
 import {
   FILE_SYSTEM_ROUTES_COMPONENTS_DIR,
   FILE_SYSTEM_ROUTES_DYNAMIC_REGEXP,
   FILE_SYSTEM_ROUTES_INDEX,
   FILE_SYSTEM_ROUTES_LAYOUT,
-} from '../constants';
+} from '../../constants';
 import { replaceWithAlias } from '../utils';
+import { makeLegalIdentifier } from '../makeLegalIdentifier';
 import { debug, findLayout, shouldSkip, getRouteWeight } from './utils';
+
+export type { RouteLegacy as Route };
 
 const compName = (srcDirectory: string, filePath: string) => {
   const legalCompName = makeLegalIdentifier(
@@ -28,10 +26,10 @@ const layoutNameAbbr = (filePath: string) => {
   return `${prefix}${makeLegalIdentifier(dirName)}`;
 };
 
-const parents: PageRoute[] = [];
+const parents: RouteLegacy[] = [];
 
 /* eslint-disable no-param-reassign */
-const recursiveReadDir = ({
+const recursiveReadDirLegacy = ({
   dir,
   routes,
   basePath = '/',
@@ -39,7 +37,7 @@ const recursiveReadDir = ({
   srcAlias,
 }: {
   dir: string;
-  routes: PageRoute[];
+  routes: RouteLegacy[];
   basePath: string;
   srcDirectory: string;
   srcAlias: string;
@@ -56,19 +54,19 @@ const recursiveReadDir = ({
     } else {
       const alias = replaceWithAlias(srcDirectory, layout, srcAlias);
       const componentName = compName(srcDirectory, layout);
-      const route: PageRoute = {
+      const route: RouteLegacy = {
         path: `${basePath.substring(0, basePath.length - 1)}`,
-        children: [],
+        exact: false,
+        routes: [],
         _component: alias,
         component: componentName,
         parent,
-        type: 'page',
       };
       parent = route;
       resetParent = true;
       routes.push(route);
       parents.push(route);
-      routes = route.children as PageRoute[];
+      routes = route.routes as RouteLegacy[];
     }
   }
 
@@ -92,7 +90,7 @@ const recursiveReadDir = ({
         }
       }
 
-      const route: PageRoute = {
+      const route: RouteLegacy = {
         path: `${basePath}${
           dynamicRouteMatched
             ? `:${dynamicRouteMatched[1].replace(/\$$/, '?')}${
@@ -102,12 +100,12 @@ const recursiveReadDir = ({
         }`,
         _component: alias,
         component: componentName,
+        exact: true,
         parent,
-        type: 'page',
       };
 
       if (fs.statSync(filePath).isDirectory()) {
-        recursiveReadDir({
+        recursiveReadDirLegacy({
           dir: filePath,
           routes,
           basePath: `${route.path}/`,
@@ -130,6 +128,7 @@ const recursiveReadDir = ({
 
       if (filename === '404' && basePath === '/') {
         route.path = '*';
+        route.exact = false;
       }
 
       routes.push(route);
@@ -143,21 +142,19 @@ const recursiveReadDir = ({
 /* eslint-enable  no-param-reassign */
 
 const normalizeNestedRoutes = (
-  nested: PageRoute[],
+  nested: RouteLegacy[],
   internalComponentsDir: string,
   internalDirectory: string,
   internalDirAlias: string,
-): PageRoute[] => {
-  const flat = (routes: PageRoute[]): PageRoute[] =>
-    routes.reduce<PageRoute[]>(
+) => {
+  const flat = (routes: RouteLegacy[]): RouteLegacy[] =>
+    routes.reduce<RouteLegacy[]>(
       (memo, route) =>
-        memo.concat(
-          Array.isArray(route.children) ? flat(route.children) : [route],
-        ),
+        memo.concat(Array.isArray(route.routes) ? flat(route.routes) : [route]),
       [],
     );
 
-  const generate = (route: PageRoute) => {
+  const generate = (route: RouteLegacy) => {
     const codes = [];
 
     let lastComponent = route.component;
@@ -217,32 +214,25 @@ export const getClientRoutes = ({
   srcAlias: string;
   internalDirectory: string;
   internalDirAlias: string;
-}): (NestedRouteForCli | PageRoute)[] => {
-  const { entryName, pageRoutesEntry } = entrypoint;
-  if (!pageRoutesEntry) {
-    return [];
-  }
-  if (!fs.existsSync(pageRoutesEntry)) {
+}) => {
+  const { entry, entryName } = entrypoint;
+
+  if (!fs.existsSync(entry)) {
     throw new Error(
-      `generate file system routes error, ${pageRoutesEntry} directory not found.`,
+      `generate file system routes error, ${entry} directory not found.`,
     );
   }
 
-  if (
-    !(
-      fs.existsSync(pageRoutesEntry) &&
-      fs.statSync(pageRoutesEntry).isDirectory()
-    )
-  ) {
+  if (!(fs.existsSync(entry) && fs.statSync(entry).isDirectory())) {
     throw new Error(
-      `generate file system routes error, ${pageRoutesEntry} should be directory.`,
+      `generate file system routes error, ${entry} should be directory.`,
     );
   }
 
-  let routes: PageRoute[] = [];
+  let routes: RouteLegacy[] = [];
 
-  recursiveReadDir({
-    dir: pageRoutesEntry,
+  recursiveReadDirLegacy({
+    dir: entry,
     routes,
     basePath: '/',
     srcDirectory,
@@ -267,10 +257,10 @@ export const getClientRoutes = ({
 
   // FIXME: support more situations
   routes.sort((a, b) => {
-    const delta = getRouteWeight(a.path!) - getRouteWeight(b.path!);
+    const delta = getRouteWeight(a.path) - getRouteWeight(b.path);
 
     if (delta === 0) {
-      return a.path!.length - b.path!.length;
+      return a.path.length - b.path.length;
     }
 
     return delta;
