+++ packages/runtime/plugin-runtime/src/router/cli/code/getClientRoutes/index.ts
