+++ packages/runtime/plugin-runtime/src/router/cli/index.ts
@@ -1,11 +1,14 @@
+import path from 'path';
 import {
   getEntryOptions,
   createRuntimeExportsUtils,
   isRouterV5 as isV5,
 } from '@modern-js/utils';
+import { cloneDeep } from '@modern-js/utils/lodash';
 import { ServerRoute } from '@modern-js/types';
 import type { CliPlugin, AppTools } from '@modern-js/app-tools';
 import { isRouteEntry, modifyEntrypoints } from './entry';
+import { isPageComponentFile } from './code/utils';
 
 export { isRouteEntry, modifyEntrypoints } from './entry';
 
@@ -13,13 +16,14 @@ const PLUGIN_IDENTIFIER = 'router';
 
 const ROUTES_IDENTIFIER = 'routes';
 
-export const routerPlugin = (): CliPlugin<AppTools> => ({
+export const routerPlugin = (): CliPlugin<AppTools<'shared'>> => ({
   name: '@modern-js/plugin-router',
   required: ['@modern-js/runtime'],
   setup: api => {
     const runtimeConfigMap = new Map<string, any>();
 
     let pluginsExportsUtils: any;
+    let originEntrypoints: any[] = [];
 
     return {
       checkEntryPoint({ path, entry }) {
@@ -50,7 +54,14 @@ export const routerPlugin = (): CliPlugin<AppTools> => ({
         };
       },
       async modifyEntrypoints({ entrypoints }) {
-        return { entrypoints: modifyEntrypoints(entrypoints) };
+        const newEntryPoints = modifyEntrypoints(entrypoints);
+        const appContext = api.useAppContext();
+        const resolvedConfig = api.useResolvedConfigContext();
+        appContext.entrypoints = newEntryPoints;
+        originEntrypoints = cloneDeep(newEntryPoints);
+        const { generateCode } = await import('./code');
+        await generateCode(appContext, resolvedConfig, entrypoints, api);
+        return { entrypoints: newEntryPoints };
       },
       modifyEntryImports({ entrypoint, imports }: any) {
         const { entryName, isMainEntry, fileSystemRoutes } = entrypoint;
@@ -121,6 +132,35 @@ export const routerPlugin = (): CliPlugin<AppTools> => ({
           pluginsExportsUtils.addExport(
             `export { default as router } from '@modern-js/runtime/router'`,
           );
+        }
+      },
+      async fileChange(e) {
+        const appContext = api.useAppContext();
+        const { appDirectory, entrypoints } = appContext;
+        const { filename, eventType } = e;
+        const nestedRouteEntries = entrypoints
+          .map(point => point.nestedRoutesEntry)
+          .filter(Boolean) as string[];
+        const pagesDir = entrypoints
+          .map(point => point.entry)
+          // should only watch file-based routes
+          .filter(entry => entry && !path.extname(entry))
+          .concat(nestedRouteEntries);
+        const isPageFile = (name: string) =>
+          pagesDir.some(pageDir => name.includes(pageDir));
+
+        const absoluteFilePath = path.resolve(appDirectory, filename);
+        const isRouteComponent =
+          isPageFile(absoluteFilePath) && isPageComponentFile(absoluteFilePath);
+
+        if (
+          isRouteComponent &&
+          (eventType === 'add' || eventType === 'unlink')
+        ) {
+          const resolvedConfig = api.useResolvedConfigContext();
+          const { generateCode } = await import('./code');
+          const entrypoints = cloneDeep(originEntrypoints);
+          await generateCode(appContext, resolvedConfig, entrypoints, api);
         }
       },
     };
